Backend/requirements.txt
asgiref==3.8.1
boto3==1.35.97
botocore==1.35.97
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==4.2
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.11.2
django-rest-framework==0.1.0
django-storages==1.14.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environ==1.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
whitenoise==6.8.2
